draw some snacks
